refactor: moved utils folder to common directory
