+++ apps/backend/src/auth/guards/local-auth.guard.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { LocalAuthGuard } from './local-auth.guard';
 import { Reflector } from '@nestjs/core';
 import { DTO_KEY } from '../../common/decorators/set-dto.decorator';
-import * as validateDtoModule from '../../utils/validation/dto-validation.util';
+import * as validateDtoModule from '../../common/utils/validation/dto-validation.util';
 import { AuthGuard, PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
