+++ apps/backend/src/common/utils/validation/environment-validation.schema.ts
